Save merged branches in reverse order
`git` returns the list of merged branches with the newest first.
AWS needs the sort order as a key, but that means re-writing
the order each time a new branch is merged. This commit reverses
the order of the list, making the oldest branch use key 0, so
that for future merged branches the keys will stay the same
and will no longer create duplicate entries.